XML for GoogleVoice Takeout
This is an XML version I added to make this compatible with the SMS Backup and Restore tools since I was having problems creating databases and interacting with the SQLLite on my android and python environments.
 
Before I talk about the work here though, I should also note that I was running this from inside Python and not from the terminal, so I modified some code outside of here. That shouldn't matter to this contribution, but I am listing some problems I've been having with this and when debugging you never know how one seemingly small change in one place will have dramatic effects that you didn't realize, so I wanted to point that out. 

In the version I downloaded, he was changing records variable with the ExplodeTextRecords, causing the need to re-run the entire program again. That wasn't easy to work with in a test environment so I created a copy of records (in this version its inside the function, but previous versions had it outside) mainly for testing. Because records is a very important array though, I wouldn't want to change its contents until we know we're done with it. (I think) I understand the need to create the TextConversationList from the records, but dealing with the amount of time it took to create records, I thought it was a better idea to modify a copy of records or a new variable instead of records itself. 

I also need to better automate the number of sms messages because (I think) that plays a large part in the tool recognizing the records. My initial assumption was that the number of records was the number of sms messages, but because records contains lists of text messages this doesn't give an accurate count. instead of Right now I've modified it outside of the program. I have ideas to correct this but haven't played with it yet. 

I was having problems getting the if statement to correctly get the type/isinstance working so I went with a string/find thing. I don't like it and seeing it seems like horrible code, but I couldn't get the type working. 

I also have a checker because I'm still getting some problems with both numbers being the "###ME###" number. In such a case, I think (for many instances) we can check a sandwich principle (if before and after are the same number we can conclude that the middle number is the same. Other cases I'm not so sure about. Once again, I had to fix this in an outside text editor. That's where I noticed the sandwich principle - it still iterates through messages in an orderly manner by the person being communicated with. 

I've also had problems with the dates. This version adds "000" to the date variables, but that gives me a date with (I think) the correct month and day but a year of 45672. Without the "000", I'm getting dates in 1969 and 1970. That's something I'm still considering.